+++ backend/src/main/java/com/suricato/entity/Confirmation.java
@@ -1,4 +1,4 @@
-package com.suricato.model.entity;
+package com.suricato.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
@@ -7,22 +7,24 @@ import org.hibernate.annotations.CreationTimestamp;
 import java.time.LocalDateTime;
 
 @Entity
-@Table(name = "ocurrence_photo")
+@Table(name = "confirmation",
+       uniqueConstraints = @UniqueConstraint(columnNames = {"user_id", "ocurrence_id"}))
 @Getter @Setter @NoArgsConstructor @AllArgsConstructor @Builder
-public class OcurrencePhoto {
+public class Confirmation {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(nullable = false, length = 512)
-    private String url;
-
-    @CreationTimestamp
-    @Column(name = "created_at", updatable = false)
-    private LocalDateTime createdAt;
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id", nullable = false)
+    private User user;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "ocurrence_id", nullable = false)
     private Ocurrence ocurrence;
+
+    @CreationTimestamp
+    @Column(name = "confirmed_at", updatable = false)
+    private LocalDateTime confirmedAt;
 }
