+++ backend/src/main/java/com/suricato/repository/CategoryRepository.java
@@ -1,6 +1,6 @@
 package com.suricato.repository;
 
-import com.suricato.model.entity.Category;
+import com.suricato.entity.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
