+++ backend/src/main/java/com/suricato/repository/CityRepository.java
@@ -1,6 +1,6 @@
 package com.suricato.repository;
 
-import com.suricato.model.entity.City;
+import com.suricato.entity.City;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CityRepository extends JpaRepository<City, Long> {
