+++ backend/src/main/java/com/suricato/repository/ConfirmationRepository.java
@@ -1,6 +1,6 @@
 package com.suricato.repository;
 
-import com.suricato.model.entity.Confirmation;
+import com.suricato.entity.Confirmation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ConfirmationRepository extends JpaRepository<Confirmation, Long> {
